Get information about a style or list styles from a Mapbox account

See the Mapbox Styles API documentation for more information.

Usage

get_style(style_id, username, style_url = NULL, access_token = NULL)

list_styles(username, access_token = NULL)

Arguments

style_id

A style ID

username

A Mapbox username

style_url

A Mapbox style URL

access_token

A Mapbox public or secret access token; set with mb_access_token()

Value

get_style returns a list of information about your selected style. list_styles returns a data frame of information about styles from a Mapbox account