Geocode an address or place description using the Mapbox Geocoding API

Geocode an address or place description using the Mapbox Geocoding API

mb_geocode(
  search_text,
  endpoint = "mapbox.places",
  limit = 1,
  types = NULL,
  search_within = NULL,
  language = NULL,
  output = "coordinates",
  access_token = NULL
)

Arguments

search_text
The text to search, formatted as a character string. Can be an address, a location, or a description of a point of interest.

endpoint
One of 'mapbox.places' (the default) or mapbox.places-permanent. Per Mapbox's terms of service, you are only allowed to save results and perform batch geocoding with the places-permanent endpoint.

limit
How many results to return; defaults to 1 (maximum 10).

types
A vector of feature types to limit to which the search should be limited. Available options include 'country', 'region', 'postcode', 'district', 'place', 'locality', 'neighborhood', 'address', and 'poi'. If left blank, all types will be searched.

search_within
An sf object, or vector representing a bounding box of format c(min_longitude, min_latitude, max_longitude, max_latitude) used to limit search results. Defaults to NULL.

language
The user's language, which can help with interpretation of queries. Available languages are found at https://docs.mapbox.com/api/search/#language-coverage.

output
If "coordinates" (the default), returns a length-two vector of coordinates or a list of coordinates. If "sf", returns an sf object with the result geometries. If "full", returns the full response from the API.

access_token
The Mapbox access token (required); can be set with mb_access_token.