Return a static Mapbox map from a specified style

Return a static Mapbox map from a specified style

static_mapbox(
  style_id,
  username,
  overlay_sf = NULL,
  overlay_style = NULL,
  overlay_markers = NULL,
  longitude = NULL,
  latitude = NULL,
  zoom = NULL,
  width = 600,
  height = 400,
  bearing = NULL,
  pitch = NULL,
  scaling_factor = c("1x", "2x"),
  before_layer = NULL,
  access_token = NULL
)

Arguments

style_id
The style ID

username
Your Mapbox username

overlay_sf
The overlay sf object (optional). The function will convert the sf object to GeoJSON then plot over the basemap style. Spatial data that are too large will trigger an error, and should be added to the style in Mapbox Studio instead.

overlay_style
A named list of vectors pecifying how to style the sf overlay. Possible names are "stroke", "stroke-width", "stroke-opacity", "fill", and "fill-opacity". The fill and stroke color values should be specified as six-digit hex codes, and the opacity and width values should be supplied as floating-point numbers.

overlay_markers
The prepared overlay markers (optional). See the function prep_overlay_markers for more information on how to specify a marker overlay.

longitude
The longitude of the map center. If an overlay is supplied, the map will default to the extent of the overlay unless longitude, latitude, and zoom are all specified.

latitude
The latitude of the map center. If an overlay is supplied, the map will default to the extent of the overlay unless longitude, latitude, and zoom are all specified.

zoom
The map zoom. The map will infer this from the overlay unless longitude, latitude, and zoom are all specified.

width
The map width; defaults to 600px

height
The map height; defaults to 600px

bearing
The map bearing; defaults to 0

pitch
The map pitch; defaults to 0

scaling_factor
The scaling factor of the tiles; either "1x" (the default) or "2x"

before_layer
A character string that specifies where in the hierarchy of layer elements the overlay should be inserted. The overlay will be placed just above the specified layer in the given Mapbox styles.

access_token
Your Mapbox access token; can be set with mb_access_token().

Value

A pointer to an image of class "magick-image". The resulting image can be manipulated further with functions from the magick package.