Generate an .mbtiles file with tippecanoe

Generate an .mbtiles file with tippecanoe

tippecanoe(
  input,
  output,
  layer_name = NULL,
  min_zoom = NULL,
  max_zoom = NULL,
  drop_rate = NULL,
  overwrite = TRUE,
  other_options = NULL,
  keep_geojson = FALSE
)

Arguments

input
The dataset from which to generate vector tiles. Can be an sf object or GeoJSON file on disk.

output
The name of the output .mbtiles file (with .mbtiles extension). Will be saved in the current working directory.

layer_name
The name of the layer in the output .mbtiles file. If NULL, will either be a random string (if input is an sf object) or the name of the input GeoJSON file (if input is a file path).

min_zoom
The minimum zoom level for which to compute tiles.

max_zoom
The maximum zoom level for which to compute tiles. If both min_zoom and max_zoom are blank, tippecanoe will guess the best zoom levels for your data.

drop_rate
The rate at which tippecanoe will drop features as you zoom out. If NULL, tippecanoe will drop features as needed in the densest tiles to stay within Mapbox's limits.

overwrite
If TRUE, an existing .mbtiles file with the same name will be overwritten.

other_options
A character string of other options to be passed to the tippecanoe program.

keep_geojson
Whether nor not to keep the temporary CSV or GeoJSON file used to generate the tiles. Defaults to FALSE.